requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
python-multipart
twilio
gtts
httpx
pytest
elevenlabs
openai-whisper
numpy
pydantic-settings
pre-commit
black
ruff
mypy
